Use `u64` instead of `u32` for opcodes and operands.
This resulted in a small performance increase.
CM2 went from begin x1.8 faster than lua to x2 times faster,
at this point I'm still using lua as a reference point but in the
future more serious benchmarking will need to be done.

+++ main.zig
@@ -51,7 +51,7 @@ const program = [_]Instruction{
 export var stack = [_]u64{0} ** 2048;
 
 // Instruction opcode.
-const Opcode = enum(u32) {
+const Opcode = enum(u64) {
     li,
     cp,
     bge,
@@ -78,7 +78,7 @@ const ops: std.EnumArray(Opcode, Operation) = ops: {
 
 // CM2 bytecode format.
 const Instruction = struct {
-    const Data = u32;
+    const Data = u64;
 
     op: Opcode,
     d0: Data = 0,
